scripts/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
opencv-python==4.8.1.78
pytesseract==0.3.10
Pillow==10.1.0
transformers==4.35.2
torch==2.1.1
numpy==1.24.3
python-multipart==0.0.6
pymongo==4.6.0
